#47 templates no longer need to be downloaded, delete MPatterns store and download patterns, rename MGamepads into useGamepads
