+++ src/assets/patterns.json
