--- src/store/models/PatternState.ts
@@ -1,5 +1,0 @@
-import type { TPattern } from '@/models';
-
-export interface IPatternState {
-  patterns: TPattern[];
-}
