--- src/store/modules/MPatterns.const.ts
@@ -1,4 +1,0 @@
-/**
- * URL, где хранятся паттерны вибраций.
- */
-export const PATH_TO_PATTERNS = 'https://wavelovers.ru/assets/patterns.json';
