+++ src/store/types/RootState.ts
@@ -1,4 +1,7 @@
+import type { TPattern } from '@/models';
+
 export interface IRootState {
   mode: number;
   isActive: boolean;
+  patterns: TPattern[];
 }
