+++ src/store/types/index.ts
@@ -1,3 +1,2 @@
 export * from './GamepadsState';
-export * from './PatternState';
 export * from './RootState';
